Move sip directory out of PYTHONPATH
Python 3 has the concept of namespace packages and
any directory with the correct name can be imported.
In this case our sip module is inside PyQt4/5 so
import sip falls back to the sip directory containing
the .sip files and the required methods don't exist.
Moving this out of the PYTHONPATH solves this issue.